requirements.txt
# Core trading and data
tinkoff-invest>=0.2.0.0
pandas>=2.2.0
numpy>=1.26.0

# Machine learning
scikit-learn>=1.5.0
statsmodels>=0.14.0
tensorflow>=2.15.0

# Optimization
optuna>=3.6.0

# Backtesting (optional)
backtrader>=1.9.78.123

# Utils
python-dotenv>=1.0.1
pyyaml>=6.0.1
requests>=2.31.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0

# Web framework
flask>=3.0.0

# Configuration
pydantic>=2.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Code quality
flake8>=6.0.0
black>=23.0.0

# Monitoring and notifications
prometheus-client>=0.19.0
grafana-api>=1.0.3
python-telegram-bot>=20.0.0
# smtplib3>=0.4.0 (optional)

# Optimization and A/B testing
optuna>=3.6.0
scipy>=1.12.0

# Logging and monitoring
loguru>=0.7.0
